Fixed single-precision InternalPath.FindIntersection precision problem for bitmap sizes > 1500 by using double-precision for intermediate calculations.

+++ src/SixLabors.Shapes/InternalPath.cs
@@ -528,7 +528,14 @@ namespace SixLabors.Shapes
             Vector2 line2Start = target.Start;
             Vector2 line2End = target.End;
 
-            float x1, y1, x2, y2, x3, y3, x4, y4;
+            // Use double precision for the intermediate calculations, because single precision calculations
+            // easily gets over the Epsilon2 threshold for bitmap sizes larger than about 1500.
+            // This is still symptom fighting though, and probably the intersection finding algorithm
+            // should be looked over in the future (making the segments fat using epsilons doesn't truely fix the
+            // robustness problem).
+            // Future potential improvement: the precision problem will be reduced if the center of the bitmap is used as origin (0, 0),
+            // this will keep coordinates smaller and relatively precision will be larger.
+            double x1, y1, x2, y2, x3, y3, x4, y4;
             x1 = line1Start.X;
             y1 = line1Start.Y;
             x2 = line1End.X;
@@ -539,23 +546,23 @@ namespace SixLabors.Shapes
             x4 = line2End.X;
             y4 = line2End.Y;
 
-            float x12 = x1 - x2;
-            float y12 = y1 - y2;
-            float x34 = x3 - x4;
-            float y34 = y3 - y4;
-            float inter = (x12 * y34) - (y12 * x34);
+            double x12 = x1 - x2;
+            double y12 = y1 - y2;
+            double x34 = x3 - x4;
+            double y34 = y3 - y4;
+            double inter = (x12 * y34) - (y12 * x34);
 
             if (inter > -Epsilon && inter < Epsilon)
             {
                 return MaxVector;
             }
 
-            float u = (x1 * y2) - (x2 * y1);
-            float v = (x3 * y4) - (x4 * y3);
-            float x = ((x34 * u) - (x12 * v)) / inter;
-            float y = ((y34 * u) - (y12 * v)) / inter;
+            double u = (x1 * y2) - (x2 * y1);
+            double v = (x3 * y4) - (x4 * y3);
+            double x = ((x34 * u) - (x12 * v)) / inter;
+            double y = ((y34 * u) - (y12 * v)) / inter;
 
-            Vector2 point = new Vector2(x, y);
+            Vector2 point = new Vector2((float)x, (float)y);
 
             if (IsOnSegments(source, target, point))
             {
